% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_one_stage_direct.R
\name{bayes_one_stage_direct}
\alias{bayes_one_stage_direct}
\title{Calculates rejection probability for one-stage Bayesian proportion test}
\usage{
bayes_one_stage_direct(
  req_p = 0.995,
  true_p = 0.995,
  gamma = 0.8,
  n = 200,
  prior_alpha = 0.5,
  prior_beta = 0.5
)
}
\arguments{
\item{req_p}{observed count}

\item{true_p}{total count}

\item{gamma}{posterior probability threshold needed to reject, defaults to .8}

\item{n}{sample size, defaults to 200}

\item{prior_alpha}{alpha parameter of beta prior, defaults to .5 (Jeffrey's prior)}

\item{prior_beta}{beta parameter of beta prior, defaults to .5 (Jeffrey's prior)}
}
\value{
a 2-element list with df of all possible observed values and probability of rejection
}
\description{
Calculates rejection probability for one-stage Bayesian proportion test
}
\examples{
bayes_one_stage_direct(true_p = .996,
                       n = 1800,
                       gamma = .7)
}
