% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_two_stage_sim.R
\name{bayes_two_stage_sim}
\alias{bayes_two_stage_sim}
\title{Helper function for Bayesian two stage simulations}
\usage{
bayes_two_stage_sim(
  req_p = 0.995,
  true_p = 0.995,
  gamma = 0.8,
  alpha = 0.05,
  n1 = 200,
  n2 = 800,
  prior_alpha = 0.5,
  prior_beta = 0.5
)
}
\arguments{
\item{req_p}{observed count}

\item{true_p}{total count}

\item{gamma}{posterior probability threshold needed to reject, defaults to .8}

\item{n1}{first stage sample size, defaults to 200}

\item{n2}{second stage sample size, defaults to 800}

\item{prior_alpha}{alpha parameter of beta prior, defaults to .5 (Jeffrey's prior)}

\item{prior_beta}{beta parameter of beta prior, defaults to .5 (Jeffrey's prior)}
}
\value{
a df of simulated outcomes and posterior probabilities
}
\description{
Helper function for Bayesian two stage simulations
}
