% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_two_stage_sim_power_calc.R
\name{bayes_two_stage_sim_power_calc}
\alias{bayes_two_stage_sim_power_calc}
\title{Calculate rejection probability of Bayesian two stage design via simulation}
\usage{
bayes_two_stage_sim_power_calc(
  n_sim = 10000,
  req_p = 0.995,
  true_p = 0.995,
  gamma = 0.8,
  n1 = 200,
  n2 = 800,
  prior_alpha = 0.5,
  prior_beta = 0.5
)
}
\arguments{
\item{n_sim}{number of simulations, defaults to 10000}

\item{req_p}{observed count}

\item{true_p}{total count}

\item{gamma}{posterior probability threshold needed to reject, defaults to .8}

\item{n1}{first stage sample size, defaults to 200}

\item{n2}{second stage sample size, defaults to 800}

\item{prior_alpha}{alpha parameter of beta prior, defaults to .5 (Jeffrey's prior)}

\item{prior_beta}{beta parameter of beta prior, defaults to .5 (Jeffrey's prior)}
}
\value{
df containing probability of overall rejection, rejection at first stage, and average sample size
}
\description{
Calculate rejection probability of Bayesian two stage design via simulation
}
\examples{
bayes_two_stage_sim_power_calc(true_p = .995,
                               gamma = .9)
}
