% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_binomial_pmf.R
\name{beta_binomial_pmf}
\alias{beta_binomial_pmf}
\title{Probability mass function for Beta-Binomial SSRE}
\usage{
beta_binomial_pmf(m, alpha, beta, successes, failures)
}
\arguments{
\item{m}{second stage sample size}

\item{successes}{number of observed successes in first n samples}

\item{failures}{number of observed failures in first n samples}

\item{prior_alpha}{alpha parameter of beta prior, defaults to .5 (Jeffrey's prior)}

\item{prior_beta}{beta parameter of beta prior, defaults to .5 (Jeffrey's prior)}
}
\value{
a df of simulated outcomes and posterior probabilities
}
\description{
Helper function used in a modified version of Zhong & Koopmeiners (2013) sample size reestimation approach.
The original paper details a design which compares two proportions and has an accompanying specification for
the reestimation procedure based on a z-statistic to save computational time.
In the case of one proportion, we can use beta-binomial distribution directly.
}
