% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_of_detection.R
\name{calculate_lod}
\alias{calculate_lod}
\title{Limit of Detection (LoD) calculations}
\usage{
calculate_lod(x, LoB = NA, alpha = 0.05)
}
\arguments{
\item{x}{data.frame containing two columns: concentration and measurement}

\item{alpha}{defaults to .05}
}
\value{
a list containing parametric LoD, a table of counts, pooled
SD estimate, and per concentration SD estimates.
}
\description{
These calculations are based on CLSI EP17-A2 (section 5.3.3.2 covers
both parametric and nonparametric calculations)
}
\details{
According to EP17, the minimal experimental design is:
- 2 reagent lots
- 1 instrument system
- 3 days
- 4 low measurand content samples
- 2 replicates per sample

Note that multiplying all of this together gives us 48 samples.
According to EP17, the minimum number of replicates for low measurand samples is 60,
so at least one of these factors will have to be increased.

For these LoD calculations, it is assumed that variability is relatively
consistent across low measurand concentrations. If this is not the case,
a trial-and-error experimental design or precision profile approach may be
needed.
}
\examples{
calculate_lod(x = data, LoB = .15)
}
