% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clopper_pearson.R
\name{clopper_pearson}
\alias{clopper_pearson}
\title{Creates Clopper-Pearson binomial confidence intervals}
\usage{
clopper_pearson(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{observed count}

\item{n}{total count}

\item{conf.level}{defaults to .95}
}
\value{
a 2-element list with "upper" and "lower"
}
\description{
Creates Clopper-Pearson binomial confidence intervals
}
\examples{
clopper_pearson(x=84, n=85, conf.level=.95)
}
