% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwer.R
\name{fwer}
\alias{fwer}
\title{Calculates FWER for a given suite of tests, requirements, and true probabilities (assumed to be greater than the requirements)}
\usage{
fwer(requirements_df, nsims = 10000, interval_type = "cp", conf.level = 0.95)
}
\arguments{
\item{nsims}{defaults to 10000}

\item{interval_type}{defaults to "cp" (Clopper-Pearson). Can be set to "ws" (Wilson-Score)}

\item{conf.level}{confidence level of the two sided intervals}

\item{requirement_df}{df containing the following columns: tests, n, prob, requirements}
}
\description{
Calculates FWER for a given suite of tests, requirements, and true probabilities (assumed to be greater than the requirements)
}
\examples{
fake_data <- data.frame(
   	tests = c("Test 1", "Test 2", "NC"),
   	n = c(100, 50, 500),
   	prob = c(.992, .962, .023),
   	requirements = c(.99, .95, .04)
)
fwer(requirements_df=fake_data, nsims=10000, conf.level=.95)
}
