% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_failures.R
\name{max_failures}
\alias{max_failures}
\title{Calculates Max Number of Sample Failures}
\usage{
max_failures(n, prob)
}
\arguments{
\item{n}{total count}

\item{prob}{defaults to .95}
}
\value{
an int for max number of sample failures, and df containing binomial probabilities
}
\description{
Calculates Max Number of Sample Failures
}
\examples{
max_failures(n=20, prob=.0125)
}
