% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_misses.R
\name{max_misses}
\alias{max_misses}
\title{Calculates Max Number of Misses before Requirement Fails}
\usage{
max_misses(
  n,
  conf.level = 0.95,
  requirement = 0.95,
  requirement_type = "gt",
  interval_type = "cp"
)
}
\arguments{
\item{n}{total count}

\item{requirement}{defaults to .95}

\item{requirement_type}{defaults to "gt" (greater than)}

\item{interval_type}{defaults to "cp" (Clopper-Pearson). Can be set to "ws".}
}
\value{
an int for max number of misses, and df containing binomial probabilities
}
\description{
Calculates Max Number of Misses before Requirement Fails
}
\examples{
max_misses(n=20, requirement = .95)
}
