% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_neighborhood.R
\name{power_neighborhood}
\alias{power_neighborhood}
\title{Search neighborhood around a given sample size for highest power}
\usage{
power_neighborhood(neighborhood, power_config)
}
\arguments{
\item{neighborhood}{the neighborhood to search in}

\item{power_config}{a list containing the specificiations for a power calculation}
}
\value{
list containing baseline power (float) and smallest sample size in neighborhood that achieves that power
}
\description{
In a binomial setting, sometimes higher power can be found by decreasing the
sample size by a small amount.
}
\examples{
power_neighborhood(neighborhood = 10, power_config = power_config)
}
