% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_one_sample_per.R
\name{select_one_sample_per}
\alias{select_one_sample_per}
\title{Random sample one case per combination of factors}
\usage{
select_one_sample_per(.data, seed = 100, ...)
}
\arguments{
\item{.data}{a dataframe}

\item{seed}{random seed by default set to 100}

\item{...}{grouping columns (e.g. patient ID, confirmed biopsy truth status, etc.)}
}
\value{
an ungrouped dataframe containing only the randomly selected rows
}
\description{
To prevent dependence between samples when repeated measures are used.
}
\examples{
nrows <- 300
test_data <- data.frame(
	patient_ids = sort(sample(1:100, nrows, replace=TRUE)),
	test = runif(n=nrows),
	casefile_id = seq_along(1:nrows)
)

test_data \%>\% 
	mutate(
		status = if_else(test < .80, "STABLE", "AR")
	) \%>\% 
	select(-test) -> test_data

select_one_sample_per(test_data, patient_ids, status)
}
