% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_signatera_confidence.R
\name{sig_adjust_confidences}
\alias{sig_adjust_confidences}
\title{Adjust Signatera positive calls}
\usage{
sig_adjust_confidences(input_df, amplicon_id = TRUE)
}
\arguments{
\item{amplicon_id}{defaults to TRUE. Adds "ref", "mut" columns based on optional amplicon_id column}

\item{df}{containing columns for "ref", "mut", "confidence"}
}
\value{
df with "adjustedMutationPresent" column added
}
\description{
New updated threshold for C>A/G>T and C>T/G>A from 0.97→ 0.98
}
\examples{
adjust_sig_positives(df)
}
