% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_vaf_filter.R
\name{sig_constant_vaf}
\alias{sig_constant_vaf}
\title{Identify Signatera Constant VAF targets}
\usage{
sig_constant_vaf(.data, constant_vaf_bound = 0.3)
}
\arguments{
\item{.data}{data.frame containing alleleRatio and some grouping variables such as amplicon_id}

\item{constant_vaf_bound}{defaults to .3.}
}
\value{
df with a binary labelling column "constant_vaf_target": 1 if yes, 0 if no
}
\description{
Identify Signatera Constant VAF targets
}
\examples{
sig_constant_vaf(df, constant_vaf_bound = .3)
}
