% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilson_score.R
\name{wilson_score}
\alias{wilson_score}
\title{Creates Wilson Score binomial confidence intervals}
\usage{
wilson_score(x, n, conf.level = 0.95)
}
\arguments{
\item{x}{observed count}

\item{n}{total count}

\item{conf.level}{defaults to .95}
}
\value{
a 2-element list with "upper" and "lower"
}
\description{
Creates Wilson Score binomial confidence intervals
}
\examples{
wilson_score(x=84, n=85, conf.level=.95)
}
